+++ Source/Scene/CloudCollection.js
@@ -42,8 +42,7 @@ var attributeLocationsBatched = {
   positionLowAndScaleY: 1,
   packedAttribute0: 2, // show, brightness, direction
   packedAttribute1: 3, // cloudSize, slice
-  packedAttribute2: 4, // cloudColor
-
+  color: 4,
 };
 
 var attributeLocationsInstanced = {
@@ -52,7 +51,7 @@ var attributeLocationsInstanced = {
   positionLowAndScaleY: 2,
   packedAttribute0: 3, // show, brightness
   packedAttribute1: 4, // cloudSize, slice
-  packedAttribute2: 5, // cloudColor
+  color: 5,
 };
 
 var SHOW_INDEX = CumulusCloud.SHOW_INDEX;
@@ -63,7 +62,6 @@ var SLICE_INDEX = CumulusCloud.SLICE_INDEX;
 var BRIGHTNESS_INDEX = CumulusCloud.BRIGHTNESS_INDEX;
 var NUMBER_OF_PROPERTIES = CumulusCloud.NUMBER_OF_PROPERTIES;
 var COLOR_INDEX = CumulusCloud.COLOR_INDEX;
-
 
 /**
  * A renderable collection of clouds in the 3D scene.
@@ -171,9 +169,6 @@ function CloudCollection(options) {
    */
   this.noiseOffset = Cartesian3.clone(
     defaultValue(options.noiseOffset, Cartesian3.ZERO)
-  );
-  this.color = Cartesian3.clone(
-    defaultValue(options.color, Color.WHITE)
   );
 
   this._loading = false;
@@ -588,12 +583,14 @@ function createVAF(context, numberOfClouds, instanced) {
       componentsPerAttribute: 4,
       componentDatatype: ComponentDatatype.FLOAT,
       usage: BufferUsage.STATIC_DRAW,
-    },{
-      index: attributeLocations.packedAttribute2,
+    },
+    {
+      index: attributeLocations.color,
       componentsPerAttribute: 4,
-      componentDatatype: ComponentDatatype.FLOAT,
+      componentDatatype: ComponentDatatype.UNSIGNED_BYTE,
+      normalize: true,
       usage: BufferUsage.STATIC_DRAW,
-    }
+    },
   ];
 
   if (instanced) {
@@ -677,27 +674,31 @@ function writePackedAttribute1(cloudCollection, frameState, vafWriters, cloud) {
   }
 }
 
-function writePackedAttribute2(cloudCollection, frameState, vafWriters, cloud) {
+function writeColor(cloudCollection, frameState, vafWriters, cloud) {
   var i;
-  var writer = vafWriters[attributeLocations.packedAttribute2];
+  var writer = vafWriters[attributeLocations.color];
   var color = cloud.color;
-  var slice = cloud.slice;
+  var red = Color.floatToByte(color.red);
+  var green = Color.floatToByte(color.green);
+  var blue = Color.floatToByte(color.blue);
+  var alpha = Color.floatToByte(color.alpha);
+
   if (cloudCollection._instanced) {
     i = cloud._index;
-    writer(i, color.red, color.green, color.blue, slice);
+    writer(i, red, green, blue, alpha);
   } else {
     i = cloud._index * 4;
-    writer(i + 0, color.red, color.green, color.blue, slice);
-    writer(i + 1, color.red, color.green, color.blue, slice);
-    writer(i + 2, color.red, color.green, color.blue, slice);
-    writer(i + 3, color.red, color.green, color.blue, slice);
+    writer(i + 0, red, green, blue, alpha);
+    writer(i + 1, red, green, blue, alpha);
+    writer(i + 2, red, green, blue, alpha);
+    writer(i + 3, red, green, blue, alpha);
   }
 }
 function writeCloud(cloudCollection, frameState, vafWriters, cloud) {
   writePositionAndScale(cloudCollection, frameState, vafWriters, cloud);
   writePackedAttribute0(cloudCollection, frameState, vafWriters, cloud);
   writePackedAttribute1(cloudCollection, frameState, vafWriters, cloud);
-  writePackedAttribute2(cloudCollection, frameState, vafWriters, cloud);
+  writeColor(cloudCollection, frameState, vafWriters, cloud);
 }
 
 function createNoiseTexture(cloudCollection, frameState, vsSource, fsSource) {
@@ -820,8 +821,8 @@ function updateClouds(cloudCollection, frameState) {
     writers.push(writePackedAttribute1);
   }
 
-  if (properties[COLOR_INDEX] || properties[SLICE_INDEX]) {
-    writers.push(writePackedAttribute2);
+  if (properties[COLOR_INDEX]) {
+    writers.push(writeColor);
   }
 
   var numWriters = writers.length;